requirements.txt
python-telegram-bot
python-dotenv

# For weather forecast
requests

# GPT API
openai

# Local LLM
langchain_community
sentence-transformers
bs4

# VecDB
chromadb